remove grant type
